move files inside workerholic directory and require all files inside main workerholic file
