+++ lib/workerholic.rb
@@ -1,11 +1,28 @@
+$LOAD_PATH << __dir__
+
 require 'yaml'
+require 'redis'
+require 'connection_pool'
+require 'logger'
 require 'pry-byebug'
 
-require_relative 'job'
-require_relative 'worker'
-require_relative 'manager'
+require 'workerholic/manager'
+require 'workerholic/worker_balancer'
 
-require_relative '../app_test/job_test'
+require 'workerholic/job'
+require 'workerholic/job_wrapper'
 
-auto_balance = ARGV.any? { |arg| arg == '--auto-balance' }
-Workerholic::Manager.new(auto_balance: auto_balance).start
+require 'workerholic/worker'
+require 'workerholic/job_processor'
+require 'workerholic/job_scheduler'
+require 'workerholic/job_retry'
+
+require 'workerholic/storage'
+require 'workerholic/sorted_set'
+require 'workerholic/queue'
+
+require 'workerholic/job_serializer'
+require 'workerholic/statistics'
+require 'workerholic/log_manager'
+
+require_relative '../app_test/job_test' # require the application code
