+++ lib/workerholic/adapters/active_job_adapter.rb
