+++ lib/workerholic/statistics.rb
